--- a/include/mxnet/runtime/ndarray_handle.h
+++ b/include/mxnet/runtime/ndarray_handle.h
@@ -31,7 +31,7 @@ namespace mxnet {
 class NDArrayHandleObj : public Object {
  public:
   /*! \brief the Internal value. */
-  NDArray* value;
+  NDArray value;
 
   static constexpr const char* _type_key = "MXNet.NDArrayHandle";
   MXNET_DECLARE_FINAL_OBJECT_INFO(NDArrayHandleObj, Object)
@@ -41,7 +41,7 @@ class NDArrayHandle : public ObjectRef {
  public:
   explicit NDArrayHandle(NDArray* value) {
     runtime::ObjectPtr<NDArrayHandleObj> node = make_object<NDArrayHandleObj>();
-    node->value = value;
+    node->value = *value;
     data_ = std::move(node);
   }
   MXNET_DEFINE_OBJECT_REF_METHODS(NDArrayHandle, ObjectRef, NDArrayHandleObj)