--- a/include/mxnet/runtime/packed_func.h
+++ b/include/mxnet/runtime/packed_func.h
@@ -706,7 +706,8 @@ class MXNetRetValue : public MXNetPODValue_ {
   }
   MXNetRetValue& operator=(NDArrayHandle value) {
     this->SwitchToPOD(kNDArrayHandle);
-    value_.v_handle = reinterpret_cast<void*>(value->value);
+    NDArray* arr = new NDArray(value->value);
+    value_.v_handle = reinterpret_cast<void*>(arr);
     return *this;
   }
   MXNetRetValue& operator=(const PythonArg& value) {