--- a/python/mxnet/_ctypes/cached_op.py
+++ b/python/mxnet/_ctypes/cached_op.py
@@ -85,14 +85,14 @@ class CachedOp(object):
                 args = []
             type_id = default_ctx.device_typeid if default_ctx else None
             device_id = default_ctx.device_id if default_ctx else None
-            out = out if out and not isinstance(out, NDArrayBase) else (out, )
+            out_arg = out if out and not isinstance(out, NDArrayBase) else (out, )
             output_vars = _api_internal.invoke(
                 self.handle,
                 len(args),
                 *args,
                 type_id,
                 device_id,
-                *out
+                *out_arg
             )
             if out is not None:
                 return out