--- a/python/mxnet/_ctypes/cached_op.py
+++ b/python/mxnet/_ctypes/cached_op.py
@@ -73,33 +73,28 @@ class CachedOp(object):
     def __call__(self, *args, **kwargs):
         """ctypes implementation of imperative invoke wrapper"""
         # New FFI only supports numpy ndarray
-        out = kwargs.pop('out', None)
         default_ctx = kwargs.pop('default_ctx', None)
-        if kwargs:
-            raise TypeError(
-                "CachedOp.__call__ got unexpected keyword argument(s): " + \
-                ', '.join(kwargs.keys()))
         if self.is_np_sym:
             if len(args) == 1 and args[0] is None:
                 args = []
             type_id = default_ctx.device_typeid if default_ctx else None
             device_id = default_ctx.device_id if default_ctx else None
-            out_arg = out if out and not isinstance(out, NDArrayBase) else (out, )
             output_vars = _api_internal.invoke(
                 self.handle,
-                len(args),
                 *args,
                 type_id,
-                device_id,
-                *out_arg
+                device_id
             )
-            if out is not None:
-                return out
             if len(output_vars) == 1:
                 return output_vars[0]
             else:
                 return list(output_vars)
         else:
+            out = kwargs.pop('out', None)
+            if kwargs:
+                raise TypeError(
+                    "CachedOp.__call__ got unexpected keyword argument(s): " + \
+                    ', '.join(kwargs.keys()))
             if out is not None:
                 original_output = out
                 if isinstance(out, NDArrayBase):