--- a/python/mxnet/_ffi/__init__.py
+++ b/python/mxnet/_ffi/__init__.py
@@ -17,3 +17,5 @@
 """
 Acknowledgement: This file originates from incubator-tvm
 """
+from .function import _init_api, get_global_func
+from .node_generic import convert_to_node