… .github/workflows/os_x_staticbuild.yml
… python/mxnet/_ctypes/ndarray.py
… python/mxnet/_ffi/_ctypes/object.py
… python/mxnet/_ffi/_ctypes/types.py
… python/mxnet/_ffi/_cython/base.pxi
… python/mxnet/_ffi/_cython/convert.pxi
… python/mxnet/_ffi/_cython/core.pyx
… python/mxnet/_ffi/_cython/function.pxi
… python/mxnet/_ffi/_cython/object.pxi
… python/mxnet/_ffi/function.py
… python/mxnet/_ffi/node_generic.py
… python/mxnet/_ffi/object.py
… python/mxnet/container.py
… python/mxnet/cython/api.pyx
… python/mxnet/cython/base.pyi
… python/mxnet/cython/ndarray.pyx
… python/mxnet/ndarray/_internal.py
… python/mxnet/optimizer/adamW.py
… python/setup.py
… src/api/_api_internal/_api_internal.cc
… src/api/cached_op_api.cc
… src/runtime/container.cc
… src/runtime/ndarray_handle.cc
… tests/python/unittest/test_ffi_container.py