## Description ##
This is the follow up PR for RFC #19672. Map container is added and more data types are supported by new FFI, like dictionary, list of strings. 

- [x] Make ADT container and MAP container support NDArray type. 
- [x] Adopt PackedFunc based FFI on CachedOp.  
    - Some CachedOp functions are implemented: **create, free, invoke, get_optimized_symbol**

## Checklist ##
### Essentials ###
- [x] PR's title starts with a category (e.g. [BUGFIX], [MODEL], [TUTORIAL], [FEATURE], [DOC], etc)
- [ ] Changes are complete (i.e. I finished coding on this PR)
- [x] All changes have test coverage
- [x] Code is well-documented