@barry-jin 2021-01-18
@mxnet-bot run ci [unix-cpu, unix-gpu]