@barry-jin 2021-02-09
@mxnet-bot run ci [centos-gpu, windows-gpu]